backend/requirements.txt
fastapi
uvicorn
httpx
websockets
python-multipart
pydantic
python-dotenv